seoul_gas/requirements.txt
numpy
pandas
fastapi
uvicorn
supabase
pydantic
google-genai
scikit-learn
python-dotenv
python-multipart
pydantic-settings
